pyproject.toml
[tool.poetry]
name = "ai-bio-researcher"
version = "0.1.0"
description = ""
authors = ["Marta <marta.anna.kramarczyk@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
arxiv = "2.1.3"
graspologic = "3.4.1"
imgui-bundle = "1.6.2"
importlib-metadata = "8.6.1"
ipykernel = "6.29.5"
jaraco-collections = "5.1.0"
langchain-community = "0.3.18"
langchain-openai = "0.3.7"
langgraph = "0.3.1"
moderngl = "5.12.0"
nano-vectordb = "0.0.4.3"
pickleshare = "0.7.5"
pip-chill = "1.0.3"
pyglm = "2.8.0"
pymupdf = "1.25.3"
python-louvain = "0.16"
tomli = "2.0.1"
pyyaml = "^6.0.2"
lightrag-hku = { path = "external/LightRAG", develop = true }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
